Remove stage "Run gtests"
We do no longer execute the gtests in an extra stage - they will run in the hotspot tier1 too.
The extra stage stopped working on Ubuntu 16.04docker for linux x86_64. As we could not find a fix we abandon the extra run. 
The tests will still be executed in the hotspot tier1 jtreg tests.