Refactor debian_package step
The debian-package build job was building and publishing artifacts. In order
to improve support for the aarch architecture this step should be devided
into two new parts: the "Build Debian Linux Packages" stage and a
publish-debian-package job.

The "Build Debian Linux Packages" stage produces new artifacts specific for the
target architecture. In full builds this stage will be
called twice, once for linux_x86_64 and once for linux_aarch_64

The publish-debian-package job copies the artifacts created by the
build-debian-package step to the right location