WIP: refactor debian_package step
The debian-package build job was building and publishing artifacts. In order
to improve support for the aarch architecture this step should be devided
into two new jobs: build-debian-package and publish-debian-package.

The build-debian-package job produces new artifacts specific for the
target architecture. The parameter JOB_NAME tells the job for which
architecture packages should be build. In full builds this job will be
called twice, once for linux_x86_64 and once for linux_aarch_64

The publish-debian-package copies the artifacts created by the
build-debian-package job to the right location