Rename the folder 'SapMachine-Infrastructure' to 'SapMachine-infrastructure'
The main repository for SapMachine-infrastructure is
https://github.com/SAP/SapMachine-infrastructure/ . Based on git's
default behaviour for 'git clone' this repository will be cloned into a
local repository 'SapMachine-infrastructure'. But in the jenkins
environment on ci.sapmachine.io the scripts and jobs will explicitly
clone this repo in a folder 'SapMachine-Infrastructure' where only the
second 'i' is capitalized. Developer of scripts and jobs have to keep in
mind that the local folder name is different from the last part of the
git repo name.
This commit renames the folder name SapMachine-Infrastructure to
SapMachine-infrastructure everywhere where the old name is mentioned in
the main branch of mentioned repository.
One should be aware that we only update the mentioned repository and not
repositories like
https://github.wdf.sap.corp/SapMachine/SapMachine-infrastructure.git
or developer clones on github.

This commit was produced by the following command:

git grep -l SapMachine-Infrastructure | xargs sed -i '' -e
's/SapMachine-Infrastructure/SapMachine-infrastructure/g'

When this commit is checked out the following command visualizes nicely the
exact character-level changes the commit introduced:

git show --word-diff-regex=.