Add script which downloads and tests sapmachine binaries
Downloads the sapmachine binaries for the specified versions,
platforms, file-type (dmg or tar.gz). In case of tar.gz it will expand
the archive and call 'java -version' with the java binary from the
download. It checks that the 1. line of the version string contains
'openjdk version (' followed by the expected version.
Ideas for further checks are: file size, parsing also second/third line

In order to properly downlaod from artifactory this script expects that
the environment variables ARTIFACTORY_USER and ARTIFACTORY_PWD are set.